take option

+++ test/node/test.js
@@ -1,26 +1,25 @@
-const test = require('ava');
-const fs = require('fs')
-const esbuild = require('esbuild')
-const { GasPlugin } = require('../../dist');
+const test = require("ava");
+const fs = require("fs");
+const esbuild = require("esbuild");
+const { GasPlugin } = require("../../dist");
 
 test.beforeEach(() => {
-  ((path)=>{
-    if(fs.existsSync(path))
-    fs.rmSync(path, { recursive: true })
-  })('./dist')
-})
+  ((path) => {
+    if (fs.existsSync(path)) fs.rmSync(path, { recursive: true });
+  })("./dist");
+});
 
-test('declare global functions', async t => {
-  const outfilePath = 'dist/bundle.js'
+test("declare global functions", async (t) => {
+  const outfilePath = "dist/bundle.js";
 
   await esbuild.build({
-    entryPoints: ['../fixtures/main.ts'],
+    entryPoints: ["../fixtures/main.ts"],
     bundle: true,
     outfile: outfilePath,
-    plugins: [GasPlugin]
-  })
+    plugins: [GasPlugin()],
+  });
 
-  const outfile = fs.readFileSync(outfilePath, { encoding: 'utf8' })
+  const outfile = fs.readFileSync(outfilePath, { encoding: "utf8" });
   const expected = `var global = this;
 function main1() {
 }
@@ -52,7 +51,55 @@ function main2() {
   global.main1 = main1;
   global.main2 = main2;
 })();
-`
+`;
 
-  t.is(outfile, expected)
+  t.is(outfile, expected);
 });
+
+test("option", async (t) => {
+  const outfilePath = "dist/bundle.js";
+
+  await esbuild.build({
+    entryPoints: ["../fixtures/main.ts"],
+    bundle: true,
+    outfile: outfilePath,
+    plugins: [GasPlugin({ parser: "acorn" })],
+  });
+
+  const outfile = fs.readFileSync(outfilePath, { encoding: "utf8" });
+  const expected = `var global = this;
+function main1() {
+}
+function main2() {
+}
+(() => {
+  // ../fixtures/util.ts
+  var add = (n1, n2) => n1 + n2;
+  var sub = (n1, n2) => n1 - n2;
+  var util_default = {
+    add,
+    sub
+  };
+
+  // ../fixtures/main.ts
+  var greet = (name) => {
+    console.log("Hello " + name);
+  };
+  var main1 = () => {
+    greet("mahaker");
+    console.log(util_default.add(2, 3));
+    console.log(util_default.sub(0, 5));
+  };
+  var main2 = () => {
+    greet("world!");
+    console.log(util_default.add(10, 5));
+    console.log(util_default.sub(10, 5));
+  };
+  global.main1 = main1;
+  global.main2 = main2;
+})();
+`;
+
+  t.is(outfile, expected);
+});
+
